requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.2
defusedxml==0.8.0rc2
Django==4.1.13
django-debug-toolbar==4.3.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
idna==3.6
oauthlib==3.2.2
packaging==24.0
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2023.4
redis==5.0.4
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.2
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.2.0
uWSGI==2.0.26
